set order for reshard register (#59265)
