+++ paddle/phi/core/distributed/auto_parallel/reshard/CMakeLists.txt
@@ -10,4 +10,5 @@ collect_srcs(
   s_to_s_reshard_function.cc
   p_to_s_reshard_function.cc
   nd_mesh_reshard_function.cc
-  same_status_reshard_function.cc)
+  same_status_reshard_function.cc
+  reshard_function_registry.cc)
