+++ paddle/phi/core/distributed/auto_parallel/reshard/nd_mesh_reshard_function.cc
@@ -266,7 +266,5 @@ void SameNdMeshReshardFunction::Eval(phi::DeviceContext* dev_ctx,
   }
 }
 
-REGISTER_RESHARD_FUNC(SameNdMeshReshardFunction);
-
 }  // namespace distributed
 }  // namespace phi
