+++ paddle/phi/core/distributed/auto_parallel/reshard/p_to_s_reshard_function.cc
@@ -89,7 +89,5 @@ void PToSReshardFunction::Eval(DeviceContext* dev_ctx,
   SetDistProps(out, in.dims(), out_dist_attr);
 }
 
-REGISTER_RESHARD_FUNC(PToSReshardFunction);
-
 }  // namespace distributed
 }  // namespace phi
